streamlit/requirements.txt
opencv-python
numpy
requests
Pillow
duckdb